requirements.txt
flask==2.3.2
requests==2.31.0
pandas==2.0.3
matplotlib==3.7.2
python-dotenv==1.0.0